+++ src/api.rs
@@ -72,6 +72,7 @@ pub mod endpoint_prelude;
 pub mod common;
 pub mod groups;
 pub mod projects;
+pub mod retry;
 pub mod users;
 
 pub(crate) mod helpers;